ComputerKurzweilPropertiesTest.java : Junit testing for
ComputerKurzweilProperties.propertiesFactory. This test check wheater it
worked or not if invaild file name used for
parameter.ComputerKurzweilPropertiesTest.java : Junit testing for
ComputerKurzweilProperties.propertiesFactory. This test check wheater it
worked or not if invaild file name used for
parameter.ComputerKurzweilPropertiesTest.java : Junit testing for
ComputerKurzweilProperties.propertiesFactory. This test check wheater it
worked or not if invaild file name used for parameter.